[Refactor] 수정 요소 로직 분리 및 RHF 분리했습니다
