+++ src/app/product/[id]/modify/_ui/ProductModifyForm/index.tsx
@@ -1,75 +1,27 @@
 "use client";
 
 import { Box, Button, Container, Heading, VStack } from "@chakra-ui/react";
-import { zodResolver } from "@hookform/resolvers/zod";
-import { FormProvider, useForm } from "react-hook-form";
-import React, { useState } from "react";
-import {
-  productModifySchema,
-  ProductModifyFormValues,
-} from "@/lib/react-hook-form/schema";
+import { FormProvider } from "react-hook-form";
+import React from "react";
 
-import usePutModifyProduct from "@/api/products/client/usePutModifyProduct";
 import ContolledInputProvider from "@/lib/react-hook-form/ContolledInputProvider";
 import ControlledNumberInput from "@/lib/react-hook-form/ControlledNumberInput";
 import { Input, Textarea } from "@chakra-ui/react";
 import ImagePreview from "@/lib/react-hook-form/ImagePreview";
-import useGetProduct from "@/api/products/client/useGetProduct";
+import useProductModifyForm from "./_hooks/useProuctModifyForm";
+import { productSchemaKey } from "@/lib/react-hook-form/productSchema";
 
 interface ProductEditFormProps {
   productId: string;
 }
 
 const ProductModifyForm = ({ productId }: ProductEditFormProps) => {
-  // 삭제된 이미지 URL들을 추적하는 state
-  const [deletedImageUrls, setDeletedImageUrls] = useState<string[]>([]);
-
-  const { data: product } = useGetProduct(productId);
-  const { name, description, stock, thumbnailUrl, detailFileUrls } =
-    product || {};
-
-  // 기존 이미지 삭제 핸들러
-  const handleRemoveExistingImage = (url: string) => {
-    setDeletedImageUrls((prev) => [...prev, url]);
-  };
-
-  const methods = useForm<ProductModifyFormValues>({
-    resolver: zodResolver(productModifySchema),
-    mode: "onChange",
-    defaultValues: {
-      name,
-      description,
-      amount: stock,
-      thumbnail: undefined,
-      detail: undefined,
-    },
-  });
-
+  const { onSubmit, methods, product } = useProductModifyForm({ productId });
+  const { thumbnailUrl, detailFileUrls } = product || {};
   const {
     handleSubmit,
     formState: { isSubmitting, isDirty, isValid },
   } = methods;
-
-  const { mutate: putModifyProduct } = usePutModifyProduct();
-
-  const onSubmit = (data: ProductModifyFormValues) => {
-    // 새 썸네일이 선택되었는지 확인
-    const hasNewThumbnail = data.thumbnail && data.thumbnail.length > 0;
-
-    // 수정 API 데이터 형식
-    const modifyApiData = {
-      productId,
-      name: data.name,
-      description: data.description,
-      stock: data.amount,
-      deletedImageIds: [], // 삭제된 이미지 ID들 (추후 구현)
-      ...(hasNewThumbnail &&
-        data.thumbnail && { newThumbnail: data.thumbnail[0] }),
-      newDetailImages: data.detail ? Array.from(data.detail) : [],
-    };
-
-    putModifyProduct(modifyApiData);
-  };
 
   return (
     <Box width={"100%"}>
@@ -88,7 +40,9 @@ const ProductModifyForm = ({ productId }: ProductEditFormProps) => {
                 </Heading>
 
                 {/* 상품명 */}
-                <ContolledInputProvider name="name" label="상품명">
+                <ContolledInputProvider
+                  name={productSchemaKey.name}
+                  label="상품명">
                   <Input
                     placeholder="상품명을 입력해주세요"
                     bg="gray.100"
@@ -97,7 +51,9 @@ const ProductModifyForm = ({ productId }: ProductEditFormProps) => {
                 </ContolledInputProvider>
 
                 {/* 상품 설명*/}
-                <ContolledInputProvider name="description" label="상품 설명">
+                <ContolledInputProvider
+                  name={productSchemaKey.description}
+                  label="상품 설명">
                   <Textarea
                     placeholder="상품 설명을 입력해주세요"
                     bg="gray.100"
@@ -106,34 +62,27 @@ const ProductModifyForm = ({ productId }: ProductEditFormProps) => {
                 </ContolledInputProvider>
 
                 {/* 재고 수량*/}
-                <ControlledNumberInput name="amount" label="재고 수량" />
+                <ControlledNumberInput
+                  name={productSchemaKey.amount}
+                  label="재고 수량"
+                />
 
                 {/* 메인 이미지 */}
                 <ImagePreview
-                  name="thumbnail"
+                  name={productSchemaKey.thumbnail}
                   label="메인 이미지 (변경하려면 새로 선택)"
                   multiple={false}
-                  existingImageUrls={
-                    thumbnailUrl && !deletedImageUrls.includes(thumbnailUrl)
-                      ? [thumbnailUrl]
-                      : []
-                  }
-                  onRemoveExistingImage={handleRemoveExistingImage}
+                  existingImageUrls={thumbnailUrl ? [thumbnailUrl] : []}
+                  deletedImageUrlsFieldName="deletedImageUrls"
                 />
 
                 {/* 상세 이미지 */}
                 <ImagePreview
-                  name="detail"
+                  name={productSchemaKey.detail}
                   label="상세 이미지 (추가하려면 선택)"
                   multiple={true}
-                  existingImageUrls={
-                    detailFileUrls
-                      ? detailFileUrls.filter(
-                          (url) => !deletedImageUrls.includes(url)
-                        )
-                      : []
-                  }
-                  onRemoveExistingImage={handleRemoveExistingImage}
+                  existingImageUrls={detailFileUrls || []}
+                  deletedImageUrlsFieldName="deletedImageUrls"
                 />
 
                 <Button
