+++ src/lib/react-hook-form/constant.ts
@@ -1,0 +1,3 @@
+export const MAX_FILE_SIZE = 5 * 1024 * 1024; // 5MB
+export const ACCEPTED_IMAGE_TYPES = ["image/jpeg"];
+export const MAX_DETAIL_IMAGES = 10;
